requirements.txt
# Core AI Framework (WITHOUT torch dependencies)
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langgraph>=0.1.0

# LLM Integrations (lightweight only)
langchain-google-genai>=1.0.0
langchain-groq>=0.1.0

# Web & Data
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# UI
streamlit>=1.35.0

# Data Processing
pandas>=2.0.0

# Utilities
python-dotenv>=1.0.0